[FEATURE] Ember/Glimmer interop
This PR adds basic Ember/Glimmer interop for GlimmerX. This allows users
to write libraries and components using GlimmerX in both frameworks. It
works by providing alternative implementations for each package, one for
standard Glimmer/TS/Webpack builds, one for Ember builds. Eventually the
goal is to unify these as much as possible, but for now this unlocks
experimentation and cross-compatiblity.

Changes include:

- Babel plugin has been updated to include an `ember` option and the
  ability to pass a custom `precompileTemplate` function, for the Ember
  side to use.
- Most Glimmer imports have been replaced with no-ops (like `on` and
  `fn`). Ember templates and components still rely on resolution, so
  these "just work" anyways.
- User "imports" will continue to work in Ember, as long as the files
  are exported in the correct place, since Ember still relies on
  resolution and doesn't have true imports.
- Services work, and an empty base `Service` class has been added so
  that users can import and extend it, like they would in Ember. In the
  Ember side, services need to be exported from the `/services` folder
  like normal. In the Glimmer side, they need to be instantiated and
  passed to `renderComponent`.

The PR also adds smoke tests for the new functionality, to make sure we
don't accidentally break it.